Stuf

+++ Engine64/Timer.h
@@ -5,9 +5,9 @@
 namespace Timer {
 
 	//Returns the time between the last call to this function and this one. Call this every frame.
-	float getDelta() {
+	double getDelta() {
 		static unsigned long long int lastTime = 0;
-		float f = float(glutGet(GLUT_ELAPSED_TIME) - lastTime) / 1000;
+		double f = double(glutGet(GLUT_ELAPSED_TIME) - lastTime) / 1000;
 		lastTime = glutGet(GLUT_ELAPSED_TIME);
 		return f;
 	}
